fix: consolidate schema migration for MTO BigInt and Deals constraints
- Altered `mto_delivery` columns (`quantity_traded`, `deliverable_qty`, `sr_no`) to `BIGINT` to prevent integer overflow.
- Updated `MTODelivery` model in `backend/ingest/nse_models.py` to match the new schema.
- Added logic to dynamically drop `UNIQUE` constraints on `bulk_deals` and `block_deals` to support duplicate entries.
- Removed redundant `backend/ingest/fix_mto_schema.py`.

+++ backend/ingest/nse_models.py
@@ -187,7 +187,7 @@ class MTODelivery(Base, TimescaleMixin):
     id = Column(Integer, autoincrement=True, nullable=False)
     trade_date = Column(Date, nullable=False, index=True)
     settlement_type = Column(String(10), default='N')
-    sr_no = Column(Integer)
+    sr_no = Column(BigInteger)
     security_name = Column(String(200), nullable=False)
     quantity_traded = Column(BigInteger)
     deliverable_qty = Column(BigInteger)
